dashboard hours display in server finished
