+++ client/src/services/hoursHelper.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 export default async function dashboardHelper(token) {
-  const response = await axios.get("/hours", {
+  const response = await axios.get("/displayhours", {
     headers: { Authorization: "Bearer " + token },
   });
   return response;
